[#22]; refactor: command 위치
